Copy initialize function for force generator
