+++ src/components/ideal/initialize_force_generator.hpp
@@ -8,13 +8,4 @@
 
 #include "force_generator.hpp"
 
-/**
- * @fn InitializeForceGenerator
- * @brief Initialize function for ForceGenerator
- * @param [in] clock_generator: Clock generator
- * @param [in] file_name: Path to initialize file
- * @param [in] dynamics: Dynamics information
- */
-ForceGenerator InitializeForceGenerator(ClockGenerator* clock_generator, const std::string file_name, const Dynamics* dynamics);
-
 #endif  // S2E_COMPONENTS_IDEAL_INITIALIZE_FORCE_GENERATOR_HPP_
